# services/starter-service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST APIs for managing starters and organizations.",
        "title": "Starter Service API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/business-units": {
            "get": {
                "description": "Retrieve business units with pagination",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Business Units"
                ],
                "summary": "List business units",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 10,
                        "description": "Page size",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/businessunit.BusinessUnitListAPIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/shared.GenericAPIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/shared.GenericAPIResponse"
                        }
                    }
                }
            }
        },
        "/business-units/{id}": {
            "get": {
                "description": "Retrieve a business unit with nested details by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Business Units"
                ],
                "summary": "Get business unit detail",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Business unit ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/businessunit.BusinessUnitDetailAPIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/shared.GenericAPIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/shared.GenericAPIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/shared.GenericAPIResponse"
                        }
                    }
                }
            }
        },
        "/departments": {
            "get": {
                "description": "Retrieve departments with optional business unit filter and pagination",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Departments"
                ],
                "summary": "List departments",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Filter by business unit ID",
                        "name": "business_unit_id",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 10,
                        "description": "Page size",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/department.DepartmentListAPIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/shared.GenericAPIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/shared.GenericAPIResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new department",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Departments"
                ],
                "summary": "Create department",
                "parameters": [
                    {
                        "description": "Department payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/department.CreateDepartmentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/department.DepartmentDetailAPIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/shared.GenericAPIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/shared.GenericAPIResponse"
                        }
                    }
                }
            }
        },
        "/departments/{id}": {
            "get": {
                "description": "Retrieve a department with nested details by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Departments"
                ],
                "summary": "Get department detail",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Department ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/department.DepartmentDetailAPIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/shared.GenericAPIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/shared.GenericAPIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/shared.GenericAPIResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a department by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Departments"
                ],
                "summary": "Delete department",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Department ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/department.DepartmentDeleteAPIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/shared.GenericAPIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/shared.GenericAPIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/shared.GenericAPIResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "Update department information by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Departments"
                ],
                "summary": "Update department",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Department ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/department.UpdateDepartmentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/department.DepartmentDetailAPIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/shared.GenericAPIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/shared.GenericAPIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/shared.GenericAPIResponse"
                        }
                    }
                }
            }
        },
        "/departments/{id}/leader": {
            "patch": {
                "description": "Assign or update the leader of a department",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Departments"
                ],
                "summary": "Assign department leader",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Department ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Leader assignment payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/department.AssignLeaderRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/department.DepartmentDetailAPIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/shared.GenericAPIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/shared.GenericAPIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/shared.GenericAPIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "businessunit.BusinessUnitDetailAPIResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/businessunit.BusinessUnitDetailResponse"
                },
                "error": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "businessunit.BusinessUnitDetailResponse": {
            "type": "object",
            "properties": {
                "company": {
                    "$ref": "#/definitions/shared.CompanyNested"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "leader": {
                    "$ref": "#/definitions/shared.LineManagerNested"
                },
                "name": {
                    "type": "string"
                },
                "shortname": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "businessunit.BusinessUnitListAPIResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/businessunit.BusinessUnitListAPIResponseData"
                },
                "error": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "businessunit.BusinessUnitListAPIResponseData": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/businessunit.BusinessUnitDetailResponse"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/httputil.RespPagination"
                }
            }
        },
        "department.AssignLeaderRequest": {
            "type": "object",
            "required": [
                "leader"
            ],
            "properties": {
                "leader": {
                    "$ref": "#/definitions/department.LeaderInfo"
                }
            }
        },
        "department.CreateDepartmentRequest": {
            "type": "object",
            "required": [
                "full_name",
                "shortname"
            ],
            "properties": {
                "business_unit_id": {
                    "type": "integer"
                },
                "full_name": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 3
                },
                "group_department_id": {
                    "type": "integer"
                },
                "leader_id": {
                    "type": "integer"
                },
                "shortname": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2
                }
            }
        },
        "department.DepartmentDeleteAPIResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/department.DepartmentDeleteResponseData"
                },
                "error": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "department.DepartmentDeleteResponseData": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "department.DepartmentDetailAPIResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/department.DepartmentDetailResponse"
                },
                "error": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "department.DepartmentListAPIResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/department.DepartmentListAPIResponseData"
                },
                "error": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "department.DepartmentListAPIResponseData": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/department.DepartmentDetailResponse"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/httputil.RespPagination"
                }
            }
        },
        "department.LeaderInfo": {
            "type": "object",
            "properties": {
                "domain": {
                    "type": "string",
                    "minLength": 1
                },
                "id": {
                    "type": "integer"
                }
            }
        },
        "department.UpdateDepartmentRequest": {
            "type": "object",
            "properties": {
                "business_unit_id": {
                    "type": "integer"
                },
                "full_name": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 3
                },
                "group_department_id": {
                    "type": "integer"
                },
                "leader_id": {
                    "type": "integer"
                },
                "shortname": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2
                }
            }
        },
...
        "shared.GenericAPIResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "error": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
    }
}